Extract cache from `yii\db\Command` to `yii\db\CacheableQueryTrait` and use it in `yii\db\Query`
